requirements.txt
apex==0.9.10dev
einops==0.3.0
numpy==1.21.5
opencv_python_headless==4.5.2.52
imgaug==0.4.0
Pillow==9.2.0
PyYAML==6.0
scikit_learn==1.1.2
scikit_video==1.1.11
scipy==1.7.3
timm==0.4.5
torch==1.8.2
torchvision==0.9.2
tqdm==4.64.0
yacs==0.1.8
